Create integration with methods of GooglePay API, they are:
- getActiveWalletID
Returns the Wallet ID of the active wallet. If there is no active wallet, a error is throw.

- getStableHardwareId
Returns the stable hardware ID of the device. Each physical Android device has a stable hardware ID which is consistent between wallets for a given device. This ID will change as a result of a factory reset.

- pushProvision
Starts the push tokenization flow in which the issuer provides most or all card details needed for Google Pay to get a valid token. Tokens added using this method are added to the active wallet.

Update README.md